pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "characterai>=1.0.1",
    "discord-py>=2.6.3",
    "python-dotenv>=1.1.1",
]
